XenEvtchn: Add join variable for stopping event channels
Add an mNeedJoin member and set it to true in stop(), then
signal the event. In the event thread, first check if mNeedJoin
is true when the Wait call returns. If it is, then return from the
thread.

This fixes a hang on the stop() path caused by the call to
std::thread::join and ensures that ring buffer associated with the
event channel (eventually) unmaps its grant references after the
event channel is stopped.